third_party/blink/web_tests/platform/generic/http/tests/security/img-crossorigin-redirect-no-cors-expected.txt

CONSOLE ERROR: Access to image at 'http://localhost:8000/security/resources/cors-redirect.php?mode=no&url=http://localhost:8000/security/resources/abe.png' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to image at 'http://localhost:8000/security/resources/cors-redirect.php?mode=no&url=http://localhost:8000/security/resources/abe-allow-star.php' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to image at 'http://localhost:8000/security/resources/cors-redirect.php?mode=no&url=http://localhost:8000/security/resources/abe-allow-credentials.php' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
CONSOLE ERROR: Access to image at 'http://localhost:8000/security/resources/cors-redirect.php?mode=no&url=http://127.0.0.1:8000/security/resources/abe-allow-credentials.php' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
Testing the handling of CORS-enabled fetch in the presence of 'No CORS' redirects.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS/FAIL descriptions are of the form, 'CORS request type': 'redirect CORS type' => 'resource'


PASS: Anonymous request: no-CORS => no-CORS image resource.
PASS: Anonymous request: no-CORS => anonymous-CORS image resource.
PASS: Credentialled request: no-CORS => credential-CORS image resource (same origin.)
PASS: Credentialled request: no-CORS => credential-CORS image resource (cross origin.)
PASS successfullyParsed is true

TEST COMPLETE
